Processamento Digital/Ciclo 1/requirements.txt
opencv-python
Pillow
scikit-image
numpy
